test: extract MockScreenshotSelectionService.swift from ScreenshotTestSupport.swift
Splits the `MockScreenshotSelectionService` test double out of
ScreenshotTestSupport.swift into its own file so that support file
holds only the `MockScreenshotCaptureService` double.

Byte-identical move of lines 40-73. Both files retain their test
imports.

ScreenshotTestSupport.swift: 73 → 38 lines. Tests: 667.
